aula 8

+++ backend/src/todos/use-cases/delete-todo.use-case.ts
@@ -1,21 +1,29 @@
-import { Injectable, Logger } from "@nestjs/common";
+import { Injectable, Logger, NotFoundException } from "@nestjs/common";
 import { DeleteTodoRepository } from "../repository";
 
 @Injectable()
 export class DeleteTodoUseCases {
-    constructor(
-        private readonly deleteTodoRepository: DeleteTodoRepository,
-        private readonly logger: Logger,
-    ) {}
+    logger: any;
+    deleteTodoRepository: any;
+    findTodoByIdRepository: any;
+
     async execute(id: string) {
-        try {
-            this.logger.log('Deleting todo...');
-            await this.deleteTodoRepository.delete(id);
-            this.logger.log('Todo deleted successfully.');
+        try{
+            this.logger.log('Deleting toDo...');
+
+     const todo = await this.findTodoByIdRepository.findById(id);
+
+     if (!todo) {
+        throw new NotFoundException('ToDo not found');
+     }
+     await this.deleteTodoRepository.delete(id);
+     this.logger.log('Todo deleted successfully.');
+          return todo;
         } catch (error) {
-            this.logger.error('Error deleting todo.', error);
+            this.logger.error(error);
             throw error;
         }
     }
 }
 
+
